requirements.txt
streamlit
crewai
supabase
line-bot-sdk
python-dotenv
requests
beautifulsoup4
pydantic
cerebras-cloud-sdk
